Initialize a client
```
$ npx create-snowpack-app packages/client \
  --template @snowpack/app-template-preact-typescript \
  --use-yarn
```

This seems like a reasonable starter template, but I'm not committed to
any of it's decisions. I'm happy to replace any or all of this client code with Gerald's spike
if we'd prefer. I just need a Snowpack client
of some sort to get the full stack tooling and deployment set up.